pyproject.toml
[project]
name = "market-research-crew"
version = "0.1.0"
description = "Market research crew for CrewAI deployment"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
  "crewai[tools]==1.12.2",
  "pydantic>=2,<3",
  "python-dotenv>=1,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=8,<9",
]

[project.scripts]
market-research-crew = "market_research_crew.main:run"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
market_research_crew = ["config/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.crewai]
type = "crew"